correct the type conversion

+++ src/api/sk_module.h
@@ -17,10 +17,10 @@ typedef struct sk_module_t {
     int         padding;
 #endif
 
-    void (*sk_module_init)();
-    int  (*sk_module_run)(struct sk_txn_t* txn);
-    int  (*sk_module_unpack)(const void* data, size_t data_len);
-    void (*sk_module_pack)(struct sk_txn_t* txn);
+    void   (*sk_module_init)();
+    int    (*sk_module_run)(struct sk_txn_t* txn);
+    size_t (*sk_module_unpack)(const void* data, size_t data_len);
+    void   (*sk_module_pack)(struct sk_txn_t* txn);
 } sk_module_t;
 
 #endif
